Correct the SCI01 compression mapping
I *think* this is the right compression mapping, it seems to work for
QfG2, but there is a *ton* of conflicting information about what
actually should be supported here. But, for now this seems to support
QfG2-SCI01 backgrounds and assets. But I'll be damned if I can find any
supporting documentation that says this is the way it's supposed to be.