extract/refactor filler implementations
start to try to implement the sierra PRNG implementation. This is
obviously not the correct implementation from the SCI0 engine. But it's
at least stable. I will need to go do some more research to track down
the correct implementation. Some scenes are *very* dependent on the
PRNG function to return the exact values its looking for otherwise
the scene will not render correctly.